Updating alert / footer / header parts
