+++ src/pages/parts/header/title-part.php
@@ -3,12 +3,12 @@
 namespace Feeds\Pages\Parts\Header;
 
 use Feeds\App;
-use Feeds\Response\Html;
+use Feeds\Response\View;
 
 App::check();
 
-/** @var Html $this */
-/** @var Header $model */
+/** @var View $this */
+/** @var Header_Model $model */
 
 ?>
 <div class="container-fluid mt-2 border-bottom">
